doc-governance-template/requirements-dev.txt
pytest>=7.4.0
pre-commit>=3.5.0
PyYAML>=6.0.1
jsonschema>=4.0.0
